add pause button
add pause button with logic, pause text
