+++ Assets/Scripts/GamePlay/GamePlayButtonsController.cs
@@ -1,0 +1,34 @@
+using System;
+using DefaultNamespace;
+using TMPro;
+using UnityEngine;
+using UnityEngine.UI;
+
+public class GamePlayButtonsController : MonoBehaviour
+{
+   [SerializeField] private Button _pauseButton;
+   [SerializeField] private TMP_Text _pauseText;
+   // private SoundEffectsPlayer _soundEffectsPlayer;
+
+   void Start()
+   {
+      // _soundEffectsPlayer = GameObject.FindGameObjectWithTag("Audio").GetComponent<SoundEffectsPlayer>();
+      _pauseButton.onClick.AddListener(OpenPauseMenu);
+
+   }
+
+   void OpenPauseMenu()
+   {
+      if (Time.timeScale == 0f)
+      {
+         Time.timeScale = 1f;
+         _pauseText.enabled = false;
+      }
+      else
+      {
+         Time.timeScale = 0f;
+         _pauseText.enabled = true;
+      }
+      // _soundEffectsPlayer.PlayClick(_soundEffectsPlayer.click);
+   }
+}
